+++ changelog.md
@@ -7,7 +7,11 @@
 - Sinigang now requires 1 Lime rather than any citrus fruit
   - Now provides Corrosion I for 0:30
 - Added Sinigang Cup (for [Miner's Delight +](https://www.curseforge.com/minecraft/mc-mods/miners-delight-plus))
-- Added recipe to craft a Cantaloupe from 4 Cantaloupe Slice
+- Matcha Latte can now be brewed in a Kettle with [Farmer's Respite](https://www.curseforge.com/minecraft/mc-mods/farmers-respite)
+- Ender Nectar is now upgraded from Matcha Latte, either in a Kettle or Cooking Pot
+- Matcha Gummy now provides Regeneration III for 0:05
+- Added recipe to craft a Mini Melon from 4 Melon Slices
+- Added recipe to craft a Cantaloupe from 4 Cantaloupe Slices
 
 ### Fixes
 - Fixed Chopped Clover recipe having an empty ingredient in some cases